+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### 17.0.0 Fixes
 
 - `[Datagrid]` Added example page to test bug when row updates on a cell with a custom component. ([#1564](https://github.com/infor-design/enterprise-ng/issues/1564))
+- `[Toolbar]` Refresh toolbar when `updated` method is called. ([#1569](https://github.com/infor-design/enterprise/issues/1569))
 
 ## 16.9.0